+++ README.md
@@ -33,8 +33,7 @@ Se você quer simplesmente baixar e dar um include, também é muito simples. Pr
 Essa é a melhor parte. Não poderia ser mais simples, veja um exemplo básico:
 
 ```php
-use OpenBoleto\Banco\BancoDoBrasil;
-use OpenBoleto\Agente;
+use OpenBoleto\Agente;use OpenBoleto\Banco\BancoDoBrasil;
 
 $sacado = new Agente('Fernando Maia', '023.434.234-34', 'ABC 302 Bloco N', '72000-000', 'Brasília', 'DF');
 $cedente = new Agente('Empresa de cosméticos LTDA', '02.123.123/0001-11', 'CLS 403 Lj 23', '71000-000', 'Brasília', 'DF');
@@ -162,7 +161,7 @@ Atualmente o OpenBoleto funciona com os bancos abaixo:
  <td>--</td>
  <td>BETA</td>
  </tr>
- 
+
  </table>
 
 ### API Banco Inter
