+++ composer.json
@@ -27,7 +27,7 @@
         "chillerlan/php-qrcode": "^5.0"
     },
     "autoload": {
-        "psr-0": {
+        "psr-4": {
             "OpenBoleto\\": "src/"
         }
     },
