+++ samples/itau.php
@@ -2,8 +2,8 @@
 
 require '../autoloader.php';
 
-use OpenBoleto\Banco\Itau;
 use OpenBoleto\Agente;
+use OpenBoleto\Banco\Itau;
 
 $sacado = new Agente('Fernando Maia', '023.434.234-34', 'ABC 302 Bloco N', '72000-000', 'Brasília', 'DF');
 $cedente = new Agente('Empresa de cosméticos LTDA', '02.123.123/0001-11', 'CLS 403 Lj 23', '71000-000', 'Brasília', 'DF');
@@ -18,7 +18,7 @@ $boleto = new Itau(array(
     'agencia' => 1724, // 4 dígitos
     'carteira' => 112, // 3 dígitos
     'conta' => 12345, // 5 dígitos
-    
+
     // Parâmetro obrigatório somente se a carteira for
     // 107, 122, 142, 143, 196 ou 198
     'codigoCliente' => 12345, // 5 dígitos
