+++ src/Exception.php
