--- src/OpenBoleto/Banco/Unicred.php
@@ -1,99 +1,0 @@
-<?php
-
-/*
- * OpenBoleto - Geração de boletos bancários em PHP
- *
- * LICENSE: The MIT License (MIT)
- *
- * Copyright (C) 2013 Estrada Virtual
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of this
- * software and associated documentation files (the "Software"), to deal in the Software
- * without restriction, including without limitation the rights to use, copy, modify,
- * merge, publish, distribute, sublicense, and/or sell copies of the Software, and to
- * permit persons to whom the Software is furnished to do so, subject to the following
- * conditions:
- *
- * The above copyright notice and this permission notice shall be included in all copies
- * or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
- * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A
- * PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
- * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
- * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
- * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- */
-
-namespace OpenBoleto\Banco;
-
-use OpenBoleto\BoletoAbstract;
-use OpenBoleto\Exception;
-
-/**
- * Classe boleto Unicred.
- *
- * @package    OpenBoleto
- * @author     Daniel Garajau <http://github.com/kriansa>
- * @copyright  Copyright (c) 2013 Estrada Virtual (http://www.estradavirtual.com.br)
- * @license    MIT License
- * @version    1.0
- */
-class Unicred extends BoletoAbstract
-{
-    /**
-     * Código do banco
-     * @var string
-     */
-    protected $codigoBanco = '136';
-
-    /**
-     * Localização do logotipo do banco, referente ao diretório de imagens
-     * @var string
-     */
-    protected $logoBanco = 'unicred.jpg';
-
-    /**
-     * Define as carteiras disponíveis para este banco
-     * @var array
-     */
-    protected $carteiras = array('11', '21', '31', '41', '51');
-
-    /**
-     * Gera o Nosso Número.
-     *
-     * @return string
-     */
-    protected function gerarNossoNumero()
-    {
-        $numero = self::zeroFill($this->getSequencial(), 10);
-        $dv = static::modulo11($numero);
-        $numero .= '-' . $dv['digito'];
-
-        return $numero;
-    }
-
-    /**
-     * Método para gerar o código da posição de 20 a 44
-     *
-     * @return string
-     * @throws \OpenBoleto\Exception
-     */
-    public function getCampoLivre()
-    {
-        $contaComDv = $this->getContaDv() !== null ? $this->getConta() . $this->getContaDv() : $this->getConta();
-        return self::zeroFill($this->getAgencia(), 4) . self::zeroFill($contaComDv, 10) . self::zeroFill($this->getNossoNumero(false), 11);
-    }
-
-    /**
-     * Retorna o campo Agência/Cedente do boleto
-     *
-     * @return string
-     */
-    public function getAgenciaCodigoCedente()
-    {
-        $contaComDv = $this->getConta() .'-'. static::getContaDv();
-        return static::zeroFill($this->getAgencia(), 4) . ' / ' . static::zeroFill($contaComDv, 10);
-    }
-}
-
